Working on deflector
So far I have made very little progress on this deflector, not sure what I'm doing and honestly online is not helping much. The best I have managed to do so far is have teh player spawn the deflector with E and put that on cooldown for 5 seconds before you can use it again. This deflector does not bounce back yet I'm working on that. Also I have a code that should despawn/delete the deflector after 2 seconds but its just not for some reason. Anyway I'm going to sleep and get back to working on it tomorrow.